Create blur effect as an extension
